Fix GoogleMaps API
Add general purpose google maps API key
Fix vision range for player

+++ server-side/src/main/java/GameModel/Player/PlayerConstants.java
@@ -11,7 +11,7 @@ public class PlayerConstants {
 	final static Integer START_GOLD = new Integer(100);
 
 	//Vision in meters
-	final static Integer START_VISION = new Integer(10);
+	final static Integer START_VISION = new Integer(250);
 
 
 	//Offline cooldown in seconds
